import more *Key notes
`recreateChain` converts a list of notes into a forest of notes, using
notes that are not replies as roots, and replies as child nodes,
recursively.

Previously, notes that are replies to notes not included in the
export, and their children, were never put in the forest, and
therefore wheren't imported.

This can be fine when importing from Twitter, since we can't really
link a note to a tweet.

It's less fine when importing from *key, because in those cases we
_can_ link to the remote note that the user was replying to.

This commit makes `recreateChain` optionally return "orphaned" note
trees, and uses that feature when importing from *key.